requirements-dev.txt
-r requirements.txt
pytest>=7.0.0
pytest-flask>=1.2.0
pytest-cov>=4.0.0
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0
mypy>=1.0.0
coverage>=7.0.0
pre-commit>=3.0.0